refactor: mover DTOs, mappers e exceções para a camada rest
