+++ src/main/java/com/sensedia/sample/consents/application/service/ConsentService.java
@@ -1,9 +1,9 @@
 package com.sensedia.sample.consents.application.service;
 
-import com.sensedia.sample.consents.dto.ConsentRequestDTO;
-import com.sensedia.sample.consents.dto.ConsentResponseDTO;
-import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
-import com.sensedia.sample.consents.dto.PageResponseDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentRequestDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentResponseDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentUpdateDTO;
+import com.sensedia.sample.consents.rest.dto.PageResponseDTO;
 
 import java.util.List;
 import java.util.UUID;
