+++ src/main/java/com/sensedia/sample/consents/application/service/impl/ConsentServiceImpl.java
@@ -3,13 +3,13 @@ package com.sensedia.sample.consents.application.service.impl;
 import com.sensedia.sample.consents.client.GitHubClient;
 import com.sensedia.sample.consents.domain.model.Consent;
 import com.sensedia.sample.consents.domain.model.ConsentHistory;
-import com.sensedia.sample.consents.dto.ConsentRequestDTO;
-import com.sensedia.sample.consents.dto.ConsentResponseDTO;
-import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
-import com.sensedia.sample.consents.dto.PageResponseDTO;
-import com.sensedia.sample.consents.exception.ConsentNotFoundException;
-import com.sensedia.sample.consents.exception.DuplicateCpfException;
-import com.sensedia.sample.consents.mapper.ConsentMapper;
+import com.sensedia.sample.consents.rest.dto.ConsentRequestDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentResponseDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentUpdateDTO;
+import com.sensedia.sample.consents.rest.dto.PageResponseDTO;
+import com.sensedia.sample.consents.rest.exception.ConsentNotFoundException;
+import com.sensedia.sample.consents.rest.exception.DuplicateCpfException;
+import com.sensedia.sample.consents.rest.mapper.ConsentMapper;
 import com.sensedia.sample.consents.domain.repository.ConsentHistoryRepository;
 import com.sensedia.sample.consents.domain.repository.ConsentRepository;
 import com.sensedia.sample.consents.application.service.ConsentService;
