+++ src/main/java/com/sensedia/sample/consents/rest/exception/DuplicateCpfException.java
@@ -1,4 +1,4 @@
-package com.sensedia.sample.consents.exception;
+package com.sensedia.sample.consents.rest.exception;
 
 public class DuplicateCpfException extends RuntimeException {
