+++ src/test/java/com/sensedia/sample/consents/integration/ConsentApiIntegrationTest.java
@@ -3,8 +3,8 @@ package com.sensedia.sample.consents.integration;
 import com.sensedia.sample.consents.ConsentsApplication;
 import com.sensedia.sample.consents.domain.model.Consent;
 import com.sensedia.sample.consents.domain.enums.ConsentStatus;
-import com.sensedia.sample.consents.dto.ConsentRequestDTO;
-import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentRequestDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentUpdateDTO;
 import com.sensedia.sample.consents.domain.repository.ConsentRepository;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
