+++ src/test/java/com/sensedia/sample/consents/service/impl/ConsentServiceImplTest.java
@@ -4,12 +4,12 @@ import com.sensedia.sample.consents.application.service.impl.ConsentServiceImpl;
 import com.sensedia.sample.consents.client.GitHubClient;
 import com.sensedia.sample.consents.domain.model.Consent;
 import com.sensedia.sample.consents.domain.enums.ConsentStatus;
-import com.sensedia.sample.consents.dto.ConsentRequestDTO;
-import com.sensedia.sample.consents.dto.ConsentResponseDTO;
-import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
-import com.sensedia.sample.consents.exception.ConsentNotFoundException;
-import com.sensedia.sample.consents.exception.DuplicateCpfException;
-import com.sensedia.sample.consents.mapper.ConsentMapper;
+import com.sensedia.sample.consents.rest.dto.ConsentRequestDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentResponseDTO;
+import com.sensedia.sample.consents.rest.dto.ConsentUpdateDTO;
+import com.sensedia.sample.consents.rest.exception.ConsentNotFoundException;
+import com.sensedia.sample.consents.rest.exception.DuplicateCpfException;
+import com.sensedia.sample.consents.rest.mapper.ConsentMapper;
 import com.sensedia.sample.consents.domain.repository.ConsentHistoryRepository;
 import com.sensedia.sample.consents.domain.repository.ConsentRepository;
 import org.junit.jupiter.api.AfterEach;
@@ -133,7 +133,7 @@ class ConsentServiceImplTest {
         when(repository.findById(id)).thenReturn(java.util.Optional.empty());
 
         assertThatThrownBy(() -> service.getConsentById(id))
-                .isInstanceOf(com.sensedia.sample.consents.exception.ConsentNotFoundException.class)
+                .isInstanceOf(ConsentNotFoundException.class)
                 .hasMessageContaining("Consentimento não encontrado");
     }
 
@@ -257,7 +257,7 @@ class ConsentServiceImplTest {
         when(repository.findById(id)).thenReturn(java.util.Optional.empty());
 
         assertThatThrownBy(() -> service.deleteConsent(id))
-                .isInstanceOf(com.sensedia.sample.consents.exception.ConsentNotFoundException.class)
+                .isInstanceOf(ConsentNotFoundException.class)
                 .hasMessageContaining("Consentimento não encontrado");
     }
